Rename files for active-row to active-item
