--- client/src/app/hooks/table-controls/active-row/index.ts
@@ -1,4 +1,0 @@
-export * from "./useActiveRowState";
-export * from "./getActiveRowDerivedState";
-export * from "./useActiveRowPropHelpers";
-export * from "./useActiveRowEffects";
