+++ client/src/app/hooks/table-controls/index.ts
@@ -9,4 +9,4 @@ export * from "./filtering";
 export * from "./sorting";
 export * from "./pagination";
 export * from "./expansion";
-export * from "./active-row";
+export * from "./active-item";
